# public/docs/openapi.yaml
openapi: 3.0.1
info:
  title: BiciMarket Payments API
  version: 1.0.0
  description: |
    **Owner**: Rocco Paoloni · **Clerk**: `payments.bicimarket` (admin-only)

    API para gestionar pagos, reembolsos, liquidaciones y payouts.

    **Importante**: Buyers y sellers no se loguean en Payments App. Las vistas administrativas usan JWT con `publicMetadata.admin=true`.
    Los endpoints server-to-server son consumidos por otras apps con `X-Service-Token`.

    **Secciones**:
    - **Pagos**: POST/GET /payments, refund, cancel, confirm (admin override)
    - **Reembolsos**: POST/GET /refunds, detail + status update (admin only)
    - **Comprobantes**: GET /receipts, POST /receipts (Buyer App), detail
    - **Liquidaciones**: GET /settlements, mark_paid (admin), detail
    - **Payouts**: POST/GET /payouts, detail + complete (admin only)
    - **Webhooks**: Mercado Pago + shipment-delivered trigger

  x-api-specs:
    base_path: /api/v1
    webhook_path: /webhooks/mercadopago
    version: v1
    next_version: v2
servers:
  - url: http://localhost:3000
    description: Local Development
    x-env: development
  - url: https://payments.bicimarket.com
    description: Production
    x-env: production
paths:
...
  /api/v1/refunds:
    get:
      summary: Listar reembolsos
      description: |
        **Listar reembolsos** con paginación y filtros.

        Admin only. Permite filtrar por paymentId, status, reason, fechas y búsqueda textual.
      tags:
        - "Reembolsos"
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: paymentId
          schema: { type: string }
          description: Filter by payment ID
        - in: query
          name: status
          schema: { type: string, enum: [pending, approved, rejected, failed] }
          description: Filter by refund status
        - in: query
          name: reason
          schema: { type: string }
          description: Filter by reason code
        - in: query
          name: from
          schema: { type: string, format: date-time }
          description: Filter by start date (ISO 8601)
        - in: query
          name: to
          schema: { type: string, format: date-time }
          description: Filter by end date (ISO 8601)
        - in: query
          name: q
          schema: { type: string }
          description: Full-text search across id, payment_id, seller_profile_id, gateway_reference
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          description: Paginated list of refunds
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Refund'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      summary: Crear reembolso (admin)
      description: |
        **Crear reembolso manualmente** desde el panel de admin.

        Requiere `Idempotency-Key`. Procesa el reembolso contra MP y actualiza el estado del pago
        (`refunded` si es total, `approved` si es parcial).
      tags:
        - "Reembolsos"
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema: { type: string, format: uuid }
          description: UUID for idempotent refund creation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                payment_id:
                  type: string
                  description: Payment ID (pay_...)
                amount_cents:
                  type: integer
                  description: Refund amount in cents
                reason:
                  type: string
                  enum: [seller_rejected, buyer_cancelled, not_delivered, manual]
                  default: manual
                  description: Reason for the refund
                seller_profile_id:
                  type: string
                  nullable: true
                  description: Seller profile ID (slp_...)
              required: [payment_id, amount_cents, reason]
      responses:
        '201':
          description: Refund created and processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Refund'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/receipts:
    get:
      summary: Listar comprobantes
      description: |
        **Listar comprobantes** (receipts) con paginación.

        Usado por Buyer App y Admin UI para mostrar "Mis comprobantes".
      tags:
        - "Comprobantes"
      security:
        - X-Service-Token: []
        - bearerAuth: []
      parameters:
        - in: query
          name: paymentId
          schema: { type: string }
          description: Filter by payment ID
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          description: List of receipts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Receipt'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      summary: Crear comprobante
      description: |
        **Crear comprobante** después de que pago sea aprobado.

        Server-to-server (Buyer App). Genera PDF y almacena URL.
      tags:
        - "Comprobantes"
      security:
        - X-Service-Token: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiptCreateRequest'
      responses:
        '201':
          description: Receipt created with PDF URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Receipt'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/payouts:
    get:
      summary: Listar payouts
      description: |
        **Listar transferencias** a sellers.

        Estados: `pending`, `in_progress`, `completed`, `failed`, `manual_review`.
      tags:
        - "Liquidaciones"
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: settlementId
          schema: { type: string }
          description: Filter by settlement ID
        - in: query
          name: status
          schema: { type: string, enum: [pending, in_progress, completed, failed, manual_review] }
        - in: query
          name: from
          schema: { type: string, format: date-time }
          description: Filter by start date (ISO 8601)
        - in: query
          name: q
          schema: { type: string }
          description: Full-text search across id, settlement_id, transfer_id
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          description: List of payouts with transfer status
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Payout'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      summary: Crear payout
      description: |
        **Crear transferencia** a Mercado Pago.

        Disparada por admin desde el panel. Transiciona settlement a `paid` y registra el payout.
        Requiere `Idempotency-Key`.
      tags:
        - "Liquidaciones"
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema: { type: string, format: uuid }
          description: UUID for idempotent payout creation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                settlement_id:
                  type: string
                  pattern: '^set_'
                  description: Settlement ID (set_01H...)
              required: [settlement_id]
      responses:
        '202':
          description: Payout initiated and in progress
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Payout'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk JWT token for UI/authenticated endpoints
    X-Service-Token:
      type: apiKey
      in: header
      name: X-Service-Token
      description: Service token for inter-app communication (Buyer, Seller, Shipping)
  schemas:
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        has_more:
          type: boolean
        next_cursor:
          type: string
          nullable: true
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              nullable: true
              description: Additional error details
...
    Refund:
      type: object
      properties:
        id:
          type: string
          description: Refund ID (ref_...)
        payment_id:
          type: string
        amount_cents:
          type: integer
        currency:
          type: string
          default: ARS
        status:
          type: string
          enum: [pending, approved, rejected, failed]
        reason:
          type: string
          description: Reason code (seller_rejected, buyer_cancelled, not_delivered, manual)
        gateway_reference:
          type: string
          nullable: true
          description: Mercado Pago refund ID
        seller_profile_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Receipt:
      type: object
      properties:
        id:
          type: string
          description: Receipt ID (rec_...)
        payment_id:
          type: string
        receipt_number:
          type: string
        receipt_url:
          type: string
          description: PDF URL
        amount_cents:
          type: integer
        currency:
          type: string
        issued_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    ReceiptCreateRequest:
      type: object
      properties:
        payment_id:
          type: string
        receipt_number:
          type: string
        receipt_url:
          type: string
        amount_cents:
          type: integer
        issued_at:
          type: string
          format: date-time
      required: [payment_id, receipt_number, receipt_url, amount_cents, issued_at]
...
    Payout:
      type: object
      properties:
        id:
          type: string
          description: Payout ID (pyt_...)
        settlement_id:
          type: string
        transfer_id:
          type: string
          nullable: true
          description: Mercado Pago transfer ID
        status:
          type: string
          enum: [pending, in_progress, completed, failed, manual_review]
        attempts:
          type: integer
        last_error:
          type: string
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Invalid request payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...